更方便的方法

+++ src/Limiter.go
@@ -3,7 +3,7 @@ package src
 import "github.com/gin-gonic/gin"
 
 // 限流装饰器
-func Limiter(cap int) func(handler gin.HandlerFunc) gin.HandlerFunc {
+func Limiter(cap int64) func(handler gin.HandlerFunc) gin.HandlerFunc {
 	limiter := NewBucket(cap, 1)
 	return func(handler gin.HandlerFunc) gin.HandlerFunc {
 		return func(ctx *gin.Context) {
